IB/mlx4: Pull IP address associated with a port once net device is discovered [issue_443536]
Upon rare timing conditions, mlx4_ib may be notified about new IP address while
mlx4_en is still busy in the 'add' function of the mlx4_interface. In this
state, mlx4_ib can't associate the IP address with a port and the event gets
dropped without updating the GID table. The solution is to update the GID table
by pulling the netdev addresses once the association with the port is known.

issue: 443536
Change-Id: Icb1a78fe4d3800a1f311522586c48e19d64a667b
Signed-off-by: Moni Shoua <monis@mellanox.com>