mlx4_ib: Fix LSO calculation [ib-next:ipoib-s1]
The current code decreases from the mss size (which is the gso_size from the
kernel skb) the size of the packet header, it shouldn't do that because the mss
that comes from the IPoIB driver already includes only the tcp payload without
the tcp/ip headers.

The result of such decrease is indication to the HW that each packet that the
HW sends is smaller than what it should be, now after the fix the HW sends the
exact amount of data.

The issue and the fix was sent by Matthew Finlay.

issue:none

Change-Id: I089a06ac48783dc7ef8057ede233b482e411eb9f
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>