IB/mlx4: Fix LSO calculation [ib-next:ipoib-s1]
The current code decreases from the mss size (which is the gso_size from
the kernel skb) the size of the packet header.
It shouldn't do that because the mss that comes from the IPoIB driver
already includes only the tcp payload without the tcp/ip headers.

The result is indication to the HW that each packet that the HW sends is
smaller than what it should be, and more packets will be sent on each
big message.

An easy way to demonstrate one more aspect of the problem is by
configuring the ipoib mtu to be less than 2*hlen (2*56) and then run
netperf with big messages.

Change-Id: Ide6913004400a7080133a2165db35867dab93749
Reported-by: Matthew Finlay <matt@mellanox.com>
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>