IB/core: Make sure that PSNs do not overflow [topic: ofed-to-ib-core]
The rq/sq->psn is 24 bits as defined in the IB spec, therefore ULPs and user
space applications shouldn't use the 8 most significant bits in the 32 bits
variables to avoid overflow in modify_qp.

Fixed the PSN usage in IPoIB and masked out the 8 most significant bits
for user-space applications.

From now on, we check in ib_modify_qp_is_ok that the PSN doesn't overflow, and
we fail if so.

issue: 416925
Change-Id: I7db9acc42564462de15bae84f4c512233fa8a62c
Signed-off-by: Majd Dibbiny <majd@mellanox.com>