admin_rdma.sh: Maintain LINUX_BASE_BRANCH file with last Linux commit.
LINUX_BASE_BRANCH will hold the commit ID from the Linux Git Tree that we
copied files from. It will be used in build script so that we can
checkout to the last supported commit.

Also added new flags:
--git-tree | --linux-tree - Path to Linux Git Tree
--linux-branch - Linux branch/commit to copy files from
--none-bare - given linux-tree is none bare git (no need to clone
it)

By default the script will checkout the Linux git a temp dir to the
needed branch. When --none-bare is given the script will copy from
the given Git without cloning anything.

Change-Id: Icdcd86249745ba3fb5bfeb1c92c3af22cdf70b22
Signed-off-by: Alaa Hleihel <alaa@mellanox.com>