IB/core: Add default GID for RoCE GID table [ib-next:roce_gid_table_mgmt]
When RoCE is used, a default GID address should be generated
for every supported RoCE type. These default GID addresses are
generated based on the IPv6 link-local address, but in contrast
to the GID based on the regular IPv6 link-local (as we generate
GID per IP address), these GIDs are also available if the net
device is down (in order to support loopback).
Moreover, these default GID addresses can't be deleted.

Change-Id: I11fa4b49f20e007d852178b427dbc2ca70e834d1
Signed-off-by: Matan Barak <matanb@mellanox.com>