IB/mlx5: IB reset flow support for kernel ULPs
This patch comes to generate simulated completions for outstanding WQEs that
weren't completed once HW was reset. It includes invoking the completion event
handler for all involved CQs, so that the ULPs will poll these CQs. We then
returned the simulated CQEs when the ULP does its polling.

This patch is necessary to implement the pci error handlers for mlx5_core.

issue: 439806
Change-Id: If444f460ade454d4e620e6ea53cb8508a5b8cd1c
Signed-off-by: Majd Dibbiny <majd@mellanox.com>