net/mlx4_core: Adjust command timeouts to conform to the firmware spec [topic: bug-fix]
The firmware spec states that the timeout for all commands should be 60 seconds.

In the past, the spec indicated that there were several classes of timeout
(short, medium, and long).  The driver has these different timeout classes.
We leave the class differentiation in the driver as-is (to protect against any
future spec changes), but set the timeout for all classes to be 60 seconds.

In addition, we fix a few commands which had hard-coded numeric timeouts specified.

issue: 492888
Change-Id: I7af89696e47d746ddbd2c019c1a387ed4d081f54
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>