IB/core: Add field 'entropy' to QP attributes [topic: roce_lag]
Entropy of QP is another attribute that is a hint, passed from the
consumer to the provider. This hint can help the provider to associate the QP
with hardware resources. For example, when working in link aggregation
mode the provider may choose a physical port according to load balancing
algorithm or with the help of an entropy value, if given.

issue: 433231
Change-Id: I4e92f69a9237bc84a751364d62116041a3833d3e
Signed-off-by: Moni Shoua <monis@mellanox.com>