net/mlx4_en: Support selective view of more than 64 statistics in ethtool [extending_counters]
The driver uses a bitmask to indicate which statistics should be
displayed to the user in ethtool.
The bitmask is u64, therefore we are limited for a selective view of up
to 64 statistics.  This patch extend the bitmap in order to show more
than 64 statistics in next patch.

In addition, add new mlx4_stats.h file and move to there statistics
structs and marcos.

Issue:418563
Issue:339647
Issue:31303
Change-Id: I69742524c9d9976710182a1e74dee45cbe43c75a
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>
Signed-off-by: Matan Barak <matanb@mellanox.com>