IB/core: Modify ib_verbs and cma in order to use roce_gid_cache
Previously, we resolved the dmac and took the smac and vlan
from the resolved address. Changing that into finding a net
device that matches the IP and vlan of the network packet
and querying the RoCE GID cache for this net device,
GID and GID type.

ocrdma driver changes were done by Somnath Kotur <Somnath.Kotur@Emulex.Com>

Change-Id: I3674aa55dc207d0b8334c5800610f3d2aca70988
Signed-off-by: Matan Barak <matanb@mellanox.com>