IB/core: Modify ib_verbs and cma in order to use roce_gid_table [ib-next:roce_gid_table_mgmt]
Previously, we resolved the dmac and took the smac and vlan
from the resolved address. Changing that into finding a net
device that matches the IP and vlan of the network packet
and querying the RoCE GID table for this net device,
GID and GID type.

ocrdma driver changes were done by Somnath Kotur <Somnath.Kotur@Emulex.Com>

Change-Id: I1579815230595fb2c1146fafeb94d4b5a687af96
Signed-off-by: Matan Barak <matanb@mellanox.com>