ofed_script/ipoibd: Add support to linux-br as a port in OVS network
If there is a port that connected to linux-br, than take from that
linux-br all its macs and add them to the OVS objects with the vlan that
we started with.

Attention: We will need to change the design, this fix is only a hack till we
re-design the daemon.

issue: none
Change-Id: Ib96d577f6d2ad6f74ff10bbb2e2e0e13af9237a6
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Alaa Hleihel <alaa@mellanox.com>