IB/cma: Add configfs for rdma_cm [net-next:roce_per_gid_net-next]
Users would like to control the behaviour of rdma_cm.
For example, old applications which doesn't set the
required RoCE gid type could be executed on RoCE V2
network types. In order to support this configuration,
we implement a configfs for rdma_cm.

In order to use the configfs, one needs to mount it and
mkdir <IB device name> inside rdma_cm directory.

The patch adds support for a single configuration file,
default_roce_mode. The mode can either be IB & RoCEv1 or
RoCEv2.

Change-Id: I517907788854aa737ce1543d132ee4f295c73a04
Signed-off-by: Matan Barak <matanb@mellanox.com>