net/mlx4: Postpone the registration of net_device [ib-next:roce_gid_table_mgmt]
The mlx4 network driver was registered in the context of the 'add'
function of the core driver (called when HW should be registered).
This makes the netdev event NETDEV_REGISTER to be sent in a context
where the answer to get_protocol_dev() callback returns NULL. This may
be confusing to listeners of netdev events.
This patch is a preparation to the patch that implements the
get_netdev() callback in the IB/mlx4 driver.

Change-Id: Id0601fc52f670272be015db3af67b67b8b293869
Signed-off-by: Moni Shoua <monis@mellanox.com>