net/mlx4: Postpone the registration of net_device [ib-next:roce_gid_mgmt_v6]
The mlx4 network driver was registered in the context of the 'add'
function of the core driver (called when HW should be registered).
This makes the netdev event NETDEV_REGISTER to be sent in a context
where the answer to get_protocol_dev() callback returns NULL. This may
be confusing to listeners of netdev events.
This patch is a preparation to the patch that implements the
get_netdev() callback in the IB/mlx4 driver.

Change-Id: Ie2455ad28883cd6d47f0de24ff54b354566d7a61
Signed-off-by: Moni Shoua <monis@mellanox.com>