net/mlx4_core: Return the admin alias GUID upon host view request [ib-next:admin_vf]
Return the admin alias GUID value upon a GET request via HOST. We do this so
that the GUID value requested by the admin is returned even if the SM has not
yet approved this GUID (e.g. the SM is down).

Note that this does not create a problem, since the virtual port will remain
down until the SM does ACK the requested GUID value.

issue: 515376
Change-Id: Ibb03b4bc828dcad0e891f6fe49bb5e3a14c0580f
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>