net/mlx4_core: Return the admin alias GUID upon host view request
Return the admin alias GUID value upon a GET request via HOST. We do this so
that the GUID value requested by the admin is returned even if the SM has not
yet approved this GUID (e.g. the SM is down).

Note that this does not create a problem, since the virtual port will remain
down until the SM does ACK the requested GUID value.

issue: 515376
Change-Id: Idb9e6306a81d90cdc5d18f600b8bb6144f158c11
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Alaa Hleihel <alaa@mellanox.com>