net/mlx4_core: Return the admin alias GUID upon host view request [ib-next:admin_vf]
Return the admin alias GUID value upon a GET request via HOST.

The value in the port may be different as SM might be down and didn't ACK for
that value. The virtual port in that case will be down till that the value
will be equal to the port value.

issue: 515376
Change-Id: Ibb03b4bc828dcad0e891f6fe49bb5e3a14c0580f
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>