IB/mlx4_ib: Alias GUID adding persistency support [ib-next:admin_vf]
In case the SM rejected an alias GUID request the PF driver shell keep trying
to acquire the specified GUID indefinitely by utilizing an exponential back
off scheme.

This should be managed per GUID entry, each entry that wasn't applied holds its
next retry information. Each record that has some work holds its next time to
run based on all its entries. The record having the lowest time will run
first when time arrives.

To let above work properly we should make sure that method to SM matches the
required values. In case record holds both set and delete values, the call must
be separated into 2 different calls.

For that we save as part of the callback's context the method and the requested
indexes to prevent a case that externally it was changed after the call so that
really the requested indexes will be evaluated.

When an GUID entry is approved we turn off its bit to have this indication.
This will prevent from re-asking it from the SM in future requests on that record
and generate redundant events (e.g. GUID change event) without any reason.

The port down event should be set only when previously it was up as done for the
port up event that set only if previously was down.

A record marked to be set when all its entries are approved, in case further
work is needed its status will be idle again.

A synchronization was added around the flows that change entries and
record state to make sure that the above process is done correctly.

issue: 515376
Change-Id: I69118dd49105021de90912c0cdf0a14befe54af3
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>