IB/mlx4: Implement ib_device callbacks [ib-next:roce_gid_table_mgmt]
get_netdev: get the net_device on the physical port of the IB transport port. In
port aggregation mode it is required to return the netdev of the active port.

modify_gid: note for a change in the RoCE gid cache. Handle this by writing to
the harsware GID table. It is possible that indexes in cahce and hardware tables
won't match so a translation is required when modifying a QP or creating an
address handle.

Change-Id: I12f46f85cc2cfa7a6ee158e7621186bf9921cf75
Signed-off-by: Moni Shoua <monis@mellanox.com>