IB/mlx4: Implement ib_device callback - modify_gid [net-next:roce_per_gid_net-next]
This is a new callbac that is required for RoCEv2 support.
In RoCE, GID table is managed in the IB core driver. The role of the
mlx4 driver is to synchronize the HW with the entries in the GID table.
Since it is possible that the same GID value will appear more than once
in the GID table (though with different attributes) it is required from
the mlx4 driver to maintain a reference counting mechanism and populate
the HW with a single value.
Since an index to the GID table is not necessarily the same as index to
the matching entry in the HW GID table, a translation between indexes is
required.

Change-Id: I3a5ca3df9b8020c96273306e5b660725c88b777a
Signed-off-by: Moni Shoua <monis@mellanox.com>