IB/ipoib: Handle QP in SQE state [ib-next:ipoib-s1]
As the result of a completion error the QP Can be transferred to SQE
states.
The driver doesn't know about that and all packets that are been sent to
that send QP  will be dropped.

The fix creates a task that after completion with error tracks the QP
state and if it is in SQE state moves it back to RTS.

Change-Id: Ia62cdb4725d942582d30c271af13b3f6e7bc4330
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>