IB/ipoib: Handle QP in SQE state [ib-next:ipoib-s1]
As the result of a completion error the QP can moved to SQE state by
the hardware. Since it's not the Error state, there are no flushes
and hence the driver doesn't know about that.

The fix creates a task that after completion with error which is not a
flush tracks the QP state and if it is in SQE state moves it back to RTS.

Change-Id: Id2be48724d5ad582cab1b72edc194397a23cd9f4
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>