IB/iser: Pass struct iser_mem_reg to iser_fast_reg_mr and iser_reg_sig_mr [ib-next:iser_for_4_1]
Instead of passing ib_sge as output variable, we pass the mem_reg
pointer to have the routines fill the rkey as well. This reduces
code duplication and extra assignments. This is a preparation step
to unify some registration logics together.

This patch does not change any functionality.

Issue: None

Change-Id: Ibf30b294a19017cbbb5f5d7d8b08d9ae7056e0f0
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Signed-off-by: Adir Lev <adirl@mellanox.com>