net/bonding: Fix potential bad memory access during bonding events [topic: fixes_for_roce_lag]
When queuing work to send the NETDEV_BONDING_INFO netdev event, it's
possible that when the work is executed, the pointer to the slave becomes
invalid. This can happen if between queuing the event and the
execution of the work, the net-device was un-ensvaled and re-enslaved . Fix that by queuing a work with the data of the slave instead of the slave structure.

Change-Id: I03760b31d1665979f3c933b43524266c82f4532d
Fixes: 69e6113343cf ('net/bonding: Notify state change on slaves')
Reported-by: Nikolay Aleksandrov <nikolay@redhat.com>
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>