IB/mlx4: Always use the correct port for mirrored multicast attachments [topic: fixes_for_roce_lag]
When attaching a QP to a multicast address in bonded mode, there was an
assumption that the port of the QP must be #1. This assumption isn't the
case under the flow which enables maximal usage of the physical ports.

Fix it by always checking the port of the original flow and create the
mirrored flow on the other port.

Change-Id: I94e811efc59c440c79531acca3b6a221072b42e8
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>