IB/mlx4: Enable send of RoCE QP1 packets with IP/UDP headers
RoCEv2 packets are sent over IP/UDP protocols.
The mlx4 driver uses a type of RAW QP to send packets for QP1 and
therefore needs to build the network headers below BTH in software.

This patch adds option to build QP1 packets with IP and UDP headers if
RoCEv2 is requested.

issue: 678076
Change-Id: I121c940e4f6270443a049605e7b582b8c587161a
Signed-off-by: Moni Shoua <monis@mellanox.com>