IB/ipoib: change init sequence ordering [ipoib-Doug-fixes]
In preparation for using per device work queues, we need to move the
start of the neighbor thread task to after ipoib_ib_dev_init and move
the destruction of the neighbor task to before ipoib_ib_dev_cleanup.
Otherwise we will end up freeing our workqueue with work possibly
still on it.

Change-Id: I3a4f4f8a303ac4d8feca46c3a7f404994ab98e3c
Signed-off-by: Doug Ledford <dledford@redhat.com>