IB/ipoib: Make ipoib_mcast_stop_thread flush the workqueue [ib-next:ipoib-Doug-fixes]
We used to pass a flush variable to mcast_stop_thread to indicate if
we should flush the workqueue or not.  This was due to some code
trying to flush a workqueue that it was currently running on which is
a no-no.  Now that we have per-device work queues, and now that
ipoib_mcast_restart_task has taken the fact that it is queued on a
single thread workqueue with all of the ipoib_mcast_join_task's and
therefore has no need to stop the join task while it runs, we can do
away with the flush parameter and unilaterally flush always.

Change-Id: If7ecd14c2c64c087e256146a898d3e8d73a28d08
Signed-off-by: Doug Ledford <dledford@redhat.com>