IB/ipoib: Handle -ENETRESET properly in our callback [ib-next:ipoib-Doug-fixes]
When the core layer calls our callback with ENETRESET as the error, we
clear the status to 0 before returning indicating that we are going to
handle the error ourselves.  This causes the core layer to not free the
mcast->mc structure, but we are releasing our reference to it by
clearing mcast->mc.  So, preserve our reference to the multicast
structure so when we next run ipoib_mcast_dev_flush, it will be able
to properly release the mcast->mc entry at the right time in the right
way.

Change-Id: Ibdd7484c68ea39673ff235b6ad781cfa09afbb1f
Signed-off-by: Doug Ledford <dledford@redhat.com>