IB/mlx4: Create and use another QP1 for RoCEv2 [net-next:roce_per_gid_net-next]
The mlx4 driver uses a special QP to implement the GSI QP. This kind of
QP allows to build the InfiniBand headers in SW to be put before the
payload that comes in with the WR. The mlx4 HW builds the packet,
calculates the ICRC and puts it at the end of the payload. This ICRC
calculation however depends on the QP configuration which is determined
when QP is modified (roce_mode during INIT->RTR). On the other hand,  ICRC
verification when packet is received does to depend on this
configuration.
Therefore, using 2 GSI QPs for send (one for each RoCE version) and 1
GSI QP for receive are required.

Change-Id: I3a8804ed5c6cdf714085d94c4f778e73a099b749
Signed-off-by: Moni Shoua <monis@mellanox.com>