IB/mlx4: Lock with RCU instead of RTNL [net-next:roce_per_gid_net-next]
The function eth_link_query_port() used to take the RTNL lock when
call to netdev_master_upper_dev_get() was necessary. This makes it
impossible to call this function with RTNL lock is held. Calling
netdev_master_upper_dev_get_rcu() and locking with RCU instead solve
this problem.

Change-Id: Iae5e9f6f92400b28b0f739112421c2b430325c1f
Signed-off-by: Moni Shoua <monis@mellanox.com>