IB/core: Add kref to IB devices [net-next:roce_per_gid_net-next]
Previously. we used device_mutex lock in order to protect
the device's list. That means that in order to guarantee a
device isn't freed while we use it, we had to lock all
devices.

Adding a kref per IB device. Before an IB device
is unregistered, we wait before its not held anymore.

Change-Id: Id5512933e2cbc5b3d31e5542e09e7c679f524d3e
Signed-off-by: Matan Barak <matanb@mellanox.com>