IB/ipoib: change init sequence ordering [ib-next:doug-v2]
In preparation for using per device work queues, we need to move the
start of the neighbor thread task to after ipoib_ib_dev_init and move
the destruction of the neighbor task to before ipoib_ib_dev_cleanup.
Otherwise we will end up freeing our workqueue with work possibly
still on it.
issue:none

Change-Id: I09eb38da6237a0bcc8970f98ebe0f348d3af840d
Signed-off-by: Doug Ledford <dledford@redhat.com>