IB/core: Add QP attribute 'flow_entropy'
The QP attribute 'flow_entropy' can be used to help the application to put QPs
into different classes for different purposes.
In RoCE LAG we use this value to put all QPs with the same flow_entropy on the
same port (instead of load balancing it). This is useful when the application
has knowledge about which port is best for a QP to achive equal balancing or
low latency for example.

issue: 515096
Change-Id: Id8ff534eeef792e1de545ffd97a32ec9ff4347f3
Signed-off-by: Moni Shoua <monis@mellanox.com>