IB/mlx4_ib: Alias GUID init to ask for active VFs
Change the init flow to ask GUIDs only for active VFs.
This is done for both SM & HOST modes so that there is no need any more
to maintain the ownership record type.

In case SM mode is used the initial value will be 0, ask the SM to assign, for the
HOST mode the initial value will be the HOST generated GUID.

This will enable out of the box experience for both probed and attached VMs.

issue: 515376
Change-Id: I24e1a64c9fe6de8204d9bb966d0d2b73f08dedfc
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Alaa Hleihel <alaa@mellanox.com>