IB/mlx4_ib: Alias GUID init to ask for active VFs [ib-next:admin_vf]
Change the init flow to ask GUIDs only for active VFs.
This is done for both SM & HOST modes so that there is no need any more
to maintain the ownership record type.

In case SM mode is used the initial value will be 0, ask the SM to assign, for the
HOST mode the initial value will be the HOST generated GUID.

This will enable out of the box experience for both probed and attached VMs.

issue: 515376
Change-Id: I5ca31c0a786f5e49b9fddf41967b4dbf9ca18347
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>