IB/mlx4_ib: Request alias GUID on demand [ib-next:admin_vf]
Request GIDs from the SM on demand, i.e., when a VF actually needs them, and
release them when the GIDs are no longer in use.

In cloud environments, this is useful for GID migrations, in which a GID is
assigned to a VF on the destination HCA, while the VF on the source HCA is shut
down (but the GID was not administratively released).

issue: 515376
Change-Id: Id4135c3ba11e24eb168176ffcf08ea90e2390ae9
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>