IB/mlx4_ib: Request alias GUID on demand
Request GIDs from the SM on demand, i.e., when a VF actually needs them, and
release them when the GIDs are no longer in use.

In cloud environments, this is useful for GID migrations, in which a GID is
assigned to a VF on the destination HCA, while the VF on the source HCA is shut
down (but the GID was not administratively released).

issue: 515376
Change-Id: I08baf98a1f64adb0967b321e7590f9501f761b91
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Alaa Hleihel <alaa@mellanox.com>