IB/core: Fix using multiples vlans over bonds
When creating a vlan over a bond, the netdevice system sends
NETDEV_CHANGEUPPER event on the bonding device. This caused us
to delete the old upper dev of bonding from our GID table.
This is fixed by only deleting the related upper device from
the GID table if the event is on the IB device's netdevice.

issue: 517234
Change-Id: I067fa86d30112eac58b308e00e07443f42809186