net/mlx4_core: Add configfs entries for setting device specific parameters [net-next:configfs]
Introducing a new mlx4_core infrastructure using the configfs to allow
set and commit specific mlx4_core low-level device configurations.

The initialization in this code creates "mlx4_core" entry in the
configfs system. After that, the user needs to use mkdir to create
configurations for specific pci device; for example "mkdir 0000:04:00.0".
This code will verify that such device exists and that
it is owned by mlx4_core.

Configfs Sub-tree Example:
-------------------------
$ tree  /sys/kernel/config/mlx4_core/
/sys/kernel/config/mlx4_core/
├── 0000:04:00.0
│   ├── commit
│   ├── dmfs_mode
│   └── ports
│       ├── 1
│       │   └── type
│       └── 2
│           └── type
└── 0000:24:00.0
    ├── commit
    ├── dmfs_mode
    └── ports
    	├── 1
    	│   └── type
    	└── 2
    	    └── type

In this patch we export the 'commit' control, a control that accepts
only a value of 1 and triggers a device restart that causes the
settings to be permanently committed.

In the next coming patches we'll export low-level device
controls of mlx4_core that will be configured through the configfs.

Issue: 397731
Change-Id: I4bfe437156a6a848c014d8e2e7dffb187944f370
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>