IB/ipoib: Add mechanism for the neigh to inform its state [ib-next:ipoib-s1]
Add callback function to the ipoib_neigh struct in order to add the
ability to inform the object that holds the neigh on its current state.

Each neigh object is kept by one and only one object (from type
path_record or multicast object), now this object can act accordingly
the change in the neigh state.

Change-Id: Icaffeb44cbc71046de9ca9e5d83d10595a39a1d9
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>