IB/ipoib: Add mechanism for internal ipoib_neigh state change notifications [ib-next:ipoib-s1]
Add callback function to the ipoib_neigh struct in order to add the
ability to inform the object that holds the neigh on its current state.

Each neigh object is kept by one and only one object (from type
path_record or multicast object), now this object can act accordingly
the change in the neigh state.

The callback should pay attention to the context it runs on, and should act/run
according to that context limitation, for example on neigh reap process the
neigh calls the callback under spinlock etc.

Change-Id: Icaffeb44cbc71046de9ca9e5d83d10595a39a1d9
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>