IB/ipoib: Add cleanup to sendonly multicast objects [ib-next:ipoib-s1]
The sendonly multicast entry created by the driver and remained in the
driver memory till the driver goes down, even if no one is going to use
that multicast entry.

It is also not part of the kernel multicast list, so even after the
ndo_set_rx_mode called it will not be removed.

The patch cleans the sendonly mcast entry by using the neigh
notification mechanism, as the following:
Each multicast entry has at least one neigh object.
The neigh object in its life cycle notifies the mcast entry that is
going to be deleted, via callback function.
The mcast callback creates a work that frees and removes that mcast
entry.

Change-Id: Ic2933a186c90d1412c465faa49db08fa562d55da
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>