cover letter [ib-next:admin_vf]
The existing implementation exposes the GID value to the VF driver. As a
result, if the SM is not reachable or incurs a delayed response, there exists a
time frame in which the VF sees an incorrect GID, i.e., not the GID that was
intended by the admin. This results in assigning an incorrect identity to the
VF. Moreover, a subsequent change in the alias GID causes a spec-incompliant
change to the VF identity. Some guest operating systems, such as Windows,
cannot tolerate such changes.

This series solves above problem by exposing the admin desired value instead of
the value that was approved by the SM. As long as the SM doesn't approve the
GID, the VF would see its link as down.

In addition, we request GIDs from the SM on demand, i.e., when a VF actually
needs them, and release them when the GIDs are no longer in use. In cloud
environments, this is useful for GID migrations, in which a GID is assigned to
a VF on the destination HCA, while the VF on the source HCA is shut down (but
the GID was not administratively released).

From compatibility reasons an explicit admin request to set/change a GUID entry
will be done immediately regardless the question whether VF is up or not. This
will let administrators to change the GUID without the need to
unbind/bind the VF.

In addition, the existing implementation doesn't support a persistency
mechanism to retry asking from the SM when an admin request has rejected. The
PF driver shall keep trying to acquire the specified GUID indefinitely by
utilizing an exponential back off scheme, this should be managed per GUID and
be aligned with other incoming admin requests.

This ability is specially needed when we moved to above scheme that asking
GUIDs on demand, we must manage the GUID's status per entry and handle
cases that some entries are temporarily rejected.

The first patch adds the persistency support and is pre-requisites for the series.
Further patches make the change to use the admin VF behavior as described above.

Finally, the default mode is changed to be HOST assigned instead of SM
assigned, this mode expects to be the typical one and doesn't depend on SM
existence as described above.

Change-Id: I96f1b1e62f4e662edbb76d2e88c61ced7da92395